fix(ai): address Gemini bot review on PR #71 (1 HIGH, 1 MEDIUM)
HIGH:
- AnthropicRefiner.Refine now rejects empty responses at the adapter
  layer: returns "anthropic refine: empty response from model %s"
  instead of surfacing a RefineOutput{Text:"", FinishReason:"stop"}
  that could silently overwrite a ticket description if the handler's
  MinOutputLen check ever regresses. Matches the empty-check already
  present in GenerateResponse.

MEDIUM:
- buildRefineUserPrompt moved out of anthropic.go into a new
  internal/ai/prompts.go. The helper is vendor-agnostic (pure string
  construction shared by all three refiner adapters Tasks 4-6);
  housing it in anthropic.go was misleading. prompts.go becomes the
  natural home for any future shared prompt infrastructure.

Tests still green, lint clean.
