+++ internal/ai/anthropic.go
@@ -119,6 +119,14 @@ func (r *AnthropicRefiner) Refine(ctx context.Context, in RefineInput) (RefineOu
 		}
 	}
 	text := strings.Join(textParts, "\n")
+	// Adapter-level empty-response guard. The CreateRound handler also
+	// rejects empty/short output via MinOutputLen (spec §3.2), so this
+	// is defense in depth — if the handler check is ever weakened, the
+	// adapter still refuses to return {Text: "", FinishReason: "stop"}
+	// that could silently overwrite a ticket description on approve.
+	if text == "" {
+		return RefineOutput{}, fmt.Errorf("anthropic refine: empty response from model %s", r.model)
+	}
 
 	inputTok := int(resp.Usage.InputTokens)
 	outputTok := int(resp.Usage.OutputTokens)
@@ -151,24 +159,4 @@ func mapAnthropicStopReason(reason anthropic.StopReason) string {
 	default:
 		return string(reason)
 	}
-}
-
-// buildRefineUserPrompt wraps the current description and optional
-// feedback in explicit delimited blocks. This is a standard prompt-
-// injection containment pattern — the model is instructed (via the
-// system prompt) to treat everything inside the blocks as input data,
-// not as instructions. Shared across all three refiner adapters so
-// injection handling is uniform.
-func buildRefineUserPrompt(currentText, feedback string) string {
-	var sb strings.Builder
-	sb.WriteString("<<<CURRENT_DESCRIPTION>>>\n")
-	sb.WriteString(currentText)
-	sb.WriteString("\n<<<END_CURRENT_DESCRIPTION>>>\n\n")
-	if feedback != "" {
-		sb.WriteString("<<<USER_FEEDBACK>>>\n")
-		sb.WriteString(feedback)
-		sb.WriteString("\n<<<END_USER_FEEDBACK>>>\n\n")
-	}
-	sb.WriteString("Refactor the description above. Return only the new description text.")
-	return sb.String()
 }
